Deploy SH - deployment script can fail

+++ deploy.sh
@@ -1,20 +1,36 @@
 #!/bin/bash
 
-echo -e '\nStarting deployment'
+echo 'Starting deployment'
 
-echo -e '\nConnecting to test-production-environment'
-ssh vagrant@192.168.50.4 '(echo -e "\nPulling image from docker"
-                          docker pull ironpeak/tictactoe
-                          str=$(docker ps -a -f "name=productiontest" | grep "productiontest")
+echo -e '\nPulling from docker'
+ssh vagrant@192.168.50.4 'docker pull ironpeak/tictactoe'
+pulldocker=$?
+if [ $pulldocker != 0 ]; then
+    echo "docker pull failed with error code $pulldocker"
+    exit $pulldocker
+fi
+
+echo -e '\nKilling and removing current'
+ssh vagrant@192.168.50.4 '(str=$(docker ps -a -f "name=productiontest" | grep "productiontest")
                           if [ ! -z "$str" ]; then
                             echo -e "\nKilling current"
                             docker kill productiontest
                             echo -e "\nRemoving old"
                             docker rm productiontest
-                          fi
-                          echo -e "\nDeploying image"
-                          docker run -p 8080:8080 -d -e NODE_ENV=production --name productiontest ironpeak/tictactoe
-                          exit)'
+                          fi)'
+killdocker=$?
+if [ $killdocker != 0 ]; then
+    echo "docker kill & remove failed with error code $killdocker"
+    exit $killdocker
+fi
+
+echo -e '\nDeploying'
+ssh vagrant@192.168.50.4 'docker run -p 8080:8080 -d -e NODE_ENV=production --name productiontest ironpeak/tictactoe'
+rundocker=$?
+if [ $rundocker != 0 ]; then
+    echo "docker run failed with error code $rundocker"
+    exit $rundocker
+fi
 
 echo -e '\nDone'
